Lab3/student/requirements.txt
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu130
torchviz
torchview
streamlit
numpy
ipykernel
wonderwords
matplotlib
plotly
altair
pandas
scikit-learn
seaborn
imblearn
transformers
datasets
evaluate